Extract map related methods to map.ts
- move code around to create commits(?
